Created token packade and moved token files
